move modals to authmodal and rename button component
