+++ components/about/ForOrgsCardContent/ForOrgsCardContent.js
@@ -1,7 +1,7 @@
 import { Row, Col } from "../../bootstrap"
 import Image from "react-bootstrap/Image"
 import styles from "./ForOrgsCardContent.module.css"
-import { SignInWithModal } from "../../auth"
+import { SignInWithButton } from "../../auth"
 import Link from "next/link"
 
 const WhyMAPLECardContent = () => (
@@ -155,7 +155,7 @@ const BenefitsCardContent = () => (
     </Row>
     <Row className="text-center mb-3">
       <Col>
-        <SignInWithModal />
+        <SignInWithButton />
       </Col>
     </Row>
   </>
