+++ components/auth/AuthModal.tsx
@@ -1,0 +1,41 @@
+import SignInModal from "./SignInModal"
+import SignUpModal from "./SignUpModal"
+import StartModal from "./StartModal"
+import ForgotPasswordModal from "./ForgotPasswordModal"
+import { AuthFlowStep, authStepChanged, useAuth } from "./redux"
+import { useAppDispatch } from "components/hooks"
+
+
+export default function AuthModal() {
+    const dispatch = useAppDispatch()
+    const { authFlowStep: currentModal } = useAuth()
+    const setCurrentModal = (step: AuthFlowStep) =>
+      dispatch(authStepChanged(step))
+    const close = () => dispatch(authStepChanged(null))
+  
+    return (
+      <span>
+        
+        <StartModal
+          show={currentModal === "start"}
+          onHide={close}
+          onSignInClick={() => setCurrentModal("signIn")}
+          onSignUpClick={() => setCurrentModal("signUp")}
+        />
+        <SignInModal
+          show={currentModal === "signIn"}
+          onHide={close}
+          onForgotPasswordClick={() => setCurrentModal("forgotPassword")}
+        />
+        <SignUpModal 
+          show={currentModal === "signUp"} 
+          onHide={close} 
+        />
+        <ForgotPasswordModal
+          show={currentModal === "forgotPassword"}
+          onHide={() => setCurrentModal("signIn")}
+        />
+      </span>
+    )
+}
+
