+++ components/auth/index.ts
@@ -1,5 +1,5 @@
 export { useAuth } from "./redux"
 export * from "./service"
-export { default as SignInWithModal } from "./SignInWithModal"
+export { default as SignInWithButton } from "./SignInWithButton"
 export { default as SignOut } from "./SignOut"
 export * from "./types"
